+++ TimetrackingTool/src/EmployeeWorktimeRegistrationForm.java
@@ -1,4 +1,4 @@
-package labor;
+
 import javax.swing.*;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.text.PlainDocument;
